bug fix, 버프 시스템 추가
- 이미 깬 퀘스트는 보상을 못 받으므로 골드 + 5000을 없애 줘야 한다.
- 스테이지 클리어하기 전에 캐릭터가 달리고 있으면 계속 달리는 버그 수정
- 캐릭터 스왑도중 메인화면/뒤로가기를 누르면 스왑했던 캐릭터가 이미지가 그대로 남아있는 버그 수정

+++ Assets/Scripts/99_Json/SkillDatas.cs
@@ -89,7 +89,7 @@
 
     public class KnightXInputData : Skillable
     {
-        public float BuffDuration;
+        public int BuffDuration;
         public float BuffSetSpeed;
         
     }
@@ -97,7 +97,7 @@
 
     public class KnightBInputData : Skillable
     {
-        public float BuffDuration;
+        public int BuffDuration;
         public int BuffIncreaseDef;
         
     }
